+++ algo/LSH_Folder/LSH.hpp
@@ -64,12 +64,12 @@ class LSH {
         Bucket*** hashtables;
     public:
         string input_file, query_file, output_file;
-        vector<vector<int>> data; // Input Data
-        vector<vector<int>> queries_data; // Query Data
+        vector<vector<double>> data; // Input Data
+        vector<vector<double>> queries_data; // Query Data
         duration<double, std::milli> ANN_time;
         duration<double, std::milli> NNB_time;
 
-        LSH(string, string, string, int L_, int N_, int k_, int R_, long long int n, int dim, vector<vector<int>>);
+        LSH(string, string, string, int L_, int N_, int k_, int R_, long long int n, int dim, vector<vector<double>>);
         ~LSH();
         int get_pointsnum() { return points_num; }
         long int get_hashtablesize() { return hashtable_size; }
@@ -84,11 +84,11 @@ class LSH {
         void set_w(int value) { w = value; }
 
         int Calculate_w();
-        vector<long long int> Specific_Hash_Value(int g, vector<int> item);
+        vector<long long int> Specific_Hash_Value(int g, vector<double> item);
 
         void print_buckets(); /* Used for Debugging */
 
-        vector<pair<long double,int>> Search_by_range2(vector<int> query,long int R_custom) ;
+        vector<pair<long double,int>> Search_by_range2(vector<double> query,long int R_custom);
 };
 
 void LSH_Insert_Points_To_Buckets(LSH* info); /* Initialize Data (Input) */
@@ -97,12 +97,12 @@ long long int mod(long long int, long int); /* This mod can handle negative valu
 
 void Print_values(); /* Used for Debugging */
 
-vector<int> Brute_by_range(vector<int> ); /* Used for Debugging */
+vector<int> Brute_by_range(vector<double> ); /* Used for Debugging */
 
 /* Finds the N-Nearest Items to the query using LSH hash function */
-vector<pair<long double, int>> Nearest_N_search(vector<int> );
+vector<pair<long double, int>> Nearest_N_search(vector<double> );
 
 /* Finds the Items that are within a range relative to the query using LSH hash function */
-vector<int> Search_by_range(vector<int> );
+vector<int> Search_by_range(vector<double> );
 
 #endif
