+++ algo/functions/functions.hpp
@@ -14,10 +14,10 @@ using namespace std;
 
 long int dim_data(); /* Get the dimensions of the data */
 long int num_of_points(); /* Get the number of the items */
-void read_file(vector<vector<int>>&, string);
+void read_file(vector<vector<double>>&, string);
 
 double Normal_distribution(); /* Generates a sequence of random normal numbers */
 
-vector <double> Nearest_N_brute(vector<vector<int>>, vector<int> , int);
+vector <double> Nearest_N_brute(vector<vector<double>>, vector<double> , int);
 
-long double euclidean_dis(vector<int> , vector<int> ); /* Calculate Euclidean Distance */
+long double euclidean_dis(vector<double> , vector<double> ); /* Calculate Euclidean Distance */
